requirement.txt
#Requirements
pandas
openpyxl
watchdog
apscheduler
python-dotenv